+++ src/it/gestioneautomobili/service/automobile/AutomobileService.java
@@ -16,7 +16,7 @@ public interface AutomobileService {
 
 	public void rimuovi(Automobile automobileInstance) throws Exception;
 
-	public int countProprietariConAutomobiliImmatricolateDopoLAnno(int annoDiImmatricolazioneInput) throws Exception;
+	public Long countProprietariConAutomobiliImmatricolateDopoLAnno(int annoDiImmatricolazioneInput) throws Exception;
 
 	public List<Automobile> listAllAutomobiliConCodFisProprietarioCheIniziaPer(String input) throws Exception;
